# frontend/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Jewelry APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "http://157.230.33.37:8080"
        }
    ],
    "paths": {
        "/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "getProductDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequestUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/invoices/invoice": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "operationId": "createSellInvoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInvoiceSellPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "operationId": "getInvoiceIndetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoice"
                ],
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/invoices/": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "operationId": "getAllInvoicesWithPagination",
                "parameters": [
                    {
                        "name": "customerName",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transactionDateFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "transactionDateEnd",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "transactionStatus",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "PENDING_PAYMENT",
                                    "PAID",
                                    "CANCELLED"
                                ]
                            }
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "Optional[DESC]",
                            "enum": [
                                "ASC",
                                "DESC"
                            ]
                        }
                    },
                    {
                        "name": "pageNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ProductPriceDto": {
                "type": "object",
                "properties": {
                    "productPriceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMetalCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "totalGemstoneCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "laborCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "markupPercentage": {
                        "maximum": 1,
                        "minimum": 0,
                        "type": "number",
                        "format": "float"
                    },
                    "salePrice": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "ProductRequestUpdate": {
                "required": [
                    "productId",
                    "productName",
                    "productPrice",
                    "productType",
                    "saleStatus",
                    "unitMeasure"
                ],
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "saleStatus": {
                        "type": "string"
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "productPrice": {
                        "$ref": "#/components/schemas/ProductPriceDto"
                    }
                }
            },
...
            "CreateInvoiceSellPayload": {
                "required": [
                    "invoiceItems",
                    "paymentMethod",
                    "transactionType"
                ],
                "type": "object",
                "properties": {
                    "transactionType": {
                        "type": "string",
                        "description": "Type of the transaction: bán ra, mua vào",
                        "example": "SELL",
                        "enum": [
                            "SELL",
                            "PURCHASE"
                        ]
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "invoiceItems": {
                        "type": "array",
                        "description": "List of items in the invoice",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceItemPayload"
                        }
                    },
                    "paymentMethod": {
                        "type": "string",
                        "description": "Method of payment",
                        "enum": [
                            "CASH",
                            "EWALLET_VNPAY"
                        ]
                    }
                }
            },
            "InvoiceItemPayload": {
                "required": [
                    "productId",
                    "unitMeasure"
                ],
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "description": "ID of the product",
                        "format": "int32"
                    },
                    "unitMeasure": {
                        "type": "string",
                        "description": "Unit measure of the product",
                        "example": "chiếc"
                    },
                    "quantity": {
                        "type": "number",
                        "description": "Quantity of the product",
                        "format": "float",
                        "example": 1
                    }
                },
                "description": "Payload for an individual invoice item"
            },
...
            "GemStone": {
                "type": "object",
                "properties": {
                    "gemstoneId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gemstoneName": {
                        "type": "string"
                    },
                    "stoneColor": {
                        "type": "string"
                    },
                    "stoneCut": {
                        "type": "string"
                    },
                    "stoneClarity": {
                        "type": "string"
                    }
                }
            },
            "MetalPriceRate": {
                "type": "object",
                "properties": {
                    "metalPriceRateId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metalTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "profitSell": {
                        "type": "number",
                        "format": "double"
                    },
                    "profitBuy": {
                        "type": "number",
                        "format": "double"
                    },
                    "exchangeRate": {
                        "type": "number",
                        "format": "double"
                    },
                    "internationalPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "metalPriceSpot": {
                        "type": "number",
                        "format": "double"
                    },
                    "sellingPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "buyingPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "effectiveDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "MetalType": {
                "type": "object",
                "properties": {
                    "metalTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metalTypeName": {
                        "type": "string"
                    },
                    "metalPurity": {
                        "type": "number",
                        "format": "double"
                    },
                    "metalGroupName": {
                        "type": "string"
                    },
                    "isAutoUpdatePrice": {
                        "type": "boolean"
                    },
                    "isOnMonitor": {
                        "type": "boolean"
                    },
                    "currentMetalPriceRate": {
                        "$ref": "#/components/schemas/MetalPriceRate"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "saleStatus": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedBy": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productMaterials": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductMaterial"
                        }
                    },
                    "productPrice": {
                        "$ref": "#/components/schemas/ProductPrice"
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ProductMaterial": {
                "type": "object",
                "properties": {
                    "productMaterialId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "materialId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gemStone": {
                        "$ref": "#/components/schemas/GemStone"
                    },
                    "metalType": {
                        "$ref": "#/components/schemas/MetalType"
                    },
                    "materialWeight": {
                        "type": "number",
                        "format": "float"
                    },
                    "materialSize": {
                        "type": "number",
                        "format": "float"
                    },
                    "metal": {
                        "type": "boolean"
                    }
                }
            },
            "ProductPrice": {
                "type": "object",
                "properties": {
                    "productPriceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMetalCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "totalGemstoneCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "laborCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "markupPercentage": {
                        "type": "number",
                        "format": "float"
                    },
                    "salePrice": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "Account": {
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "username": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "MANAGER",
                            "STAFF",
                            "CUSTOMER"
                        ]
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "phonenumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "Invoice": {
                "type": "object",
                "properties": {
                    "invoiceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "transactionDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "transactionType": {
                        "type": "string",
                        "enum": [
                            "SELL",
                            "PURCHASE"
                        ]
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "transactionClerk": {
                        "$ref": "#/components/schemas/Account"
                    },
                    "transactionClerkId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalAmount": {
                        "type": "number",
                        "format": "double"
                    },
                    "taxPercent": {
                        "type": "number",
                        "format": "float"
                    },
                    "totalDiscount": {
                        "type": "number",
                        "format": "double"
                    },
                    "totalVoucherDiscount": {
                        "type": "number",
                        "format": "double"
                    },
                    "netAmount": {
                        "type": "number",
                        "format": "double"
                    },
                    "paymentDetailsId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING_PAYMENT",
                            "PAID",
                            "CANCELLED"
                        ]
                    },
                    "invoiceItems": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceItem"
                        }
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "CASH",
                            "EWALLET_VNPAY"
                        ]
                    }
                }
            },
            "InvoiceItem": {
                "type": "object",
                "properties": {
                    "invoiceItemId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "invoiceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "componentBuy": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number",
                        "format": "float"
                    },
                    "unitPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "discountPercentage": {
                        "type": "number",
                        "format": "double"
                    },
                    "subTotal": {
                        "type": "number",
                        "format": "double"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}